src/main/resources/insightface_pytorch/requirements - 副本.txt
bcolz==1.2.1
easydict==1.10
matplotlib==3.3.4
mxnet_cu90==1.5.0
numpy==1.16.6
opencv_python==4.4.0.46
Pillow==8.3.1
scikit_learn==0.24.0
scipy==1.5.2
tensorboardX==2.2
torch==1.1.0
torchvision==0.3.0
tqdm==4.63.0